Cleaning up some viewmodel code, Line-up control
All lists and propperties are available in the Festival class. All data
is pulled from the festival class.
Festival class is modified so it can be used for the asp.net project and
C# project.
Instead of using a custom control, I am replacing that with a user
control so that as much as possible is done automatically by data
binding :-)